# api/@typespec/openapi3/openapi.yaml
openapi: 3.0.0
info:
  title: PC Store API
  version: 0.0.0
tags: []
paths:
  /api/auth/login:
    post:
      operationId: Auth_login
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/LoginResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
  /api/auth/logout:
    post:
      operationId: Auth_logout
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - type: object
                    properties:
                      message:
                        type: string
                    required:
                      - message
                  - $ref: '#/components/schemas/ErrorResponse'
  /api/auth/me:
    get:
      operationId: Auth_getMe
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/User'
                  - $ref: '#/components/schemas/ErrorResponse'
  /api/auth/register:
    post:
      operationId: Auth_register
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/User'
                  - $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
  /api/catalog/categories:
    get:
      operationId: Catalog_getCategories
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/Category'
                  - $ref: '#/components/schemas/ErrorResponse'
  /api/catalog/products:
    get:
      operationId: Catalog_getProducts
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: minPrice
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: maxPrice
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: available
          in: query
          required: false
          schema:
            type: boolean
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/ProductListResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
  /api/catalog/products/batch:
    post:
      operationId: Catalog_getProductsByIds
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  - $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
                format: int32
  /api/catalog/products/{slug}:
    get:
      operationId: Catalog_getProductBySlug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/Product'
                  - $ref: '#/components/schemas/ErrorResponse'
  /api/orders:
    post:
      operationId: Orders_createOrder
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/CreateOrderResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
    get:
      operationId: Orders_getMyOrders
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  - $ref: '#/components/schemas/ErrorResponse'
  /api/orders/{id}:
    get:
      operationId: Orders_getOrderById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/Order'
                  - $ref: '#/components/schemas/ErrorResponse'
  /api/promo:
    get:
      operationId: Promo_getPromoBlocks
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/PromoBlock'
                  - $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ApiError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details: {}
    Category:
      type: object
      required:
        - id
        - name
        - slug
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        slug:
          type: string
    CreateOrderRequest:
      type: object
      required:
        - items
        - deliveryMethod
        - recipientName
        - phone
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
                format: int32
              quantity:
                type: integer
                format: int32
            required:
              - productId
              - quantity
        deliveryMethod:
          type: string
          enum:
            - delivery
            - pickup
        recipientName:
          type: string
        phone:
          type: string
        address:
          type: string
    CreateOrderResponse:
      type: object
      required:
        - order
      properties:
        order:
          $ref: '#/components/schemas/Order'
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      required:
        - token
        - user
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
    Money:
      type: object
      required:
        - amount
      properties:
        amount:
          type: integer
          format: int32
    Order:
      type: object
      required:
        - id
        - userId
        - items
        - total
        - status
        - deliveryMethod
        - recipientName
        - phone
        - createdAt
      properties:
        id:
          type: integer
          format: int32
        userId:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        total:
          type: integer
          format: int32
        status:
          type: string
        deliveryMethod:
          type: string
        recipientName:
          type: string
        phone:
          type: string
        address:
          type: string
        createdAt:
          type: string
    OrderItem:
      type: object
      required:
        - productId
        - quantity
        - name
        - price
      properties:
        productId:
          type: integer
          format: int32
        quantity:
          type: integer
          format: int32
        name:
          type: string
        price:
          $ref: '#/components/schemas/Money'
    Product:
      type: object
      required:
        - id
        - name
        - slug
        - price
        - available
        - category
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          $ref: '#/components/schemas/Money'
        imageUrl:
          type: string
        available:
          type: boolean
        category:
          $ref: '#/components/schemas/Category'
    ProductListResponse:
      type: object
      required:
        - items
        - total
        - page
        - limit
        - totalPages
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        total:
          type: integer
          format: int32
        page:
          type: integer
          format: int32
        limit:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
    PromoBlock:
      type: object
      required:
        - id
        - title
        - text
        - product
      properties:
        id:
          type: integer
          format: int32
        title:
          type: string
        text:
          type: string
        product:
          $ref: '#/components/schemas/Product'
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
    User:
      type: object
      required:
        - id
        - email
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int32
        email:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string